Refactor main.py to utilize utility functions for output directory management and remove redundant code
